Some robots. Talky toasters (etc) now talk.

+++ src/game-world.c
@@ -515,7 +515,7 @@ static void make_noise(struct player *p)
 static void update_scent(void)
 {
 	int y, x;
-	int scent_strength[5][5] = {
+	byte scent_strength[5][5] = {
 		{2, 2, 2, 2, 2},
 		{2, 1, 1, 1, 2},
 		{2, 1, 0, 1, 2},
